==> Codigo/notebooks/01_get_molecules.qmd <==
---
title: "Connection to CHEMBL database and active molecules for Zika virus collection"
---

```{python import libraries}
# PATH TO CUSTOM MODULES
import sys

sys.path.append("../src")

# IMPORT LIBRARIES
from chembl_webresource_client.new_client import new_client
import pandas as pd

# IMPORT CUSTOM MODULES
import repository.chembl_api_connection as chembl

```

```{python paths}
# DIRECTORIES
data_path = "../data/raw/"
```

# GET MOLECULES FOR ZIKA VIRUS

```{python get zika virus target_id}
# GET THE TARGET ID WITH MAXIMUM NUMBER OF IC50 MOLECULES FOR A GIVEN ORGANISM
organism = "Zika virus"
activity = "IC50"
target_id = chembl.get_target_id_maximum_activity(organism, activity)
```

```{python get molecules for zika virus target_id}
# GET MOLECULES FOR TARGET ID AND ACTIVITY TYPE
molecules = chembl.get_molecules_from_target_activity(target_id, activity)
```

# SAVE MOLECULES TO CSV

```{python save molecules to csv}
# SAVE MOLECULES TO CSV
molecules.to_csv(data_path + target_id + "_raw.csv", index=False)
```
